Add AppImage, better argparsing and cflags handling

+++ builder/packaging.py
@@ -13,7 +13,7 @@ from pathlib import Path
 from typing import Optional
 from zipfile import ZipFile
 
-from .constants import KITHARE_VERSION, VERSION_PACKAGE_REV
+from .constants import VERSION_PACKAGE_REV
 from .downloader import ThreadedDownloader
 from .utils import BuildError, ConvertType, convert_machine, copy, rmtree, run_cmd
 
@@ -23,6 +23,8 @@ INNO_FLAGS = "/SP- /VERYSILENT /SUPPRESSMSGBOXES /NORESTART /NOCANCEL".split()
 
 INNO_COMPILER_PATH = Path("C:\\", "Program Files (x86)", "Inno Setup 6", "ISCC.exe")
 
+APPIMAGE_DOWNLOAD = "https://github.com/AppImage/AppImageKit/releases/latest/download"
+
 
 class Packager:
     """
@@ -30,14 +32,16 @@ class Packager:
     that help packaging Kithare.
     """
 
-    def __init__(self, basepath: Path, exepath: Path, machine: str):
+    def __init__(self, basepath: Path, exepath: Path, machine: str, version: str):
         """
         Initialise Packager class
         """
-        self.basepath = basepath
-        self.packaging_dir = basepath / "builder" / "packaging"
         self.exepath = exepath
         self.machine = machine
+        self.version = version
+
+        self.packaging_dir = basepath / "builder" / "packaging"
+        self.downloader: Optional[ThreadedDownloader] = None
 
     def setup(self):
         """
@@ -50,7 +54,9 @@ class Packager:
         Make portable zip package
         """
         print("Making portable binary ZIP")
-        zipname = f"kithare-{KITHARE_VERSION}-{platform.system()}-{self.machine}.zip"
+        zipname = (
+            f"kithare-{self.version}-{platform.system().lower()}-{self.machine}.zip"
+        )
         portable_zip = self.packaging_dir / "dist" / zipname
 
         portable_zip.parent.mkdir(exist_ok=True)
@@ -68,14 +74,13 @@ class WindowsPackager(Packager):
     Subclass of Packager that handles Windows packaging
     """
 
-    def __init__(self, basepath: Path, exepath: Path, machine: str):
+    def __init__(self, basepath: Path, exepath: Path, machine: str, version: str):
         """
         Initialise WindowsPackager class
         """
-        super().__init__(basepath, exepath, machine)
+        super().__init__(basepath, exepath, machine, version)
 
         self.machine = convert_machine(machine, ConvertType.WINDOWS)
-        self.downloader: Optional[ThreadedDownloader] = None
 
     def setup(self):
         """
@@ -112,9 +117,9 @@ class WindowsPackager(Packager):
 
         # Rewrite iss file, with some macros defined
         iss_file.write_text(
-            f'#define MyAppVersion "{KITHARE_VERSION}"\n'
+            f'#define MyAppVersion "{self.version}"\n'
             + f'#define MyAppArch "{self.machine}"\n'
-            + f'#define BasePath "{self.basepath.resolve()}"\n'
+            + f'#define BasePath "{self.packaging_dir.parents[1].resolve()}"\n'
             + default_iss_file.read_text()
         )
 
@@ -143,12 +148,144 @@ class LinuxPackager(Packager):
     Subclass of Packager that handles Linux packaging
     """
 
-    def __init__(self, basepath: Path, exepath: Path, machine: str, use_alien: bool):
+    def __init__(
+        self,
+        basepath: Path,
+        exepath: Path,
+        machine: str,
+        version: str,
+        use_alien: bool,
+    ):
         """
         Initialise LinuxPackager class
         """
-        super().__init__(basepath, exepath, machine)
+        super().__init__(basepath, exepath, machine, version)
+        self.version += f"-{VERSION_PACKAGE_REV}"
+
         self.use_alien = use_alien
+
+        self.appimagekitdir: Optional[Path] = None
+
+    def setup(self):
+        """
+        This prepares AppImageKit
+        """
+        try:
+            appimage_type = convert_machine(self.machine, ConvertType.APP_IMAGE)
+        except BuildError as exc:
+            print(f"Skipping AppImage generation, because {exc.emsg}\n")
+            return
+
+        self.appimagekitdir = (
+            self.packaging_dir.parents[1] / "deps" / "AppImage" / appimage_type
+        )
+
+        if self.appimagekitdir.is_dir():
+            # AppImageKit already exists, no installation required
+            return
+
+        print(
+            "Downloading and installing AppImageKit (for making AppImages) in "
+            "the background while compilation continues"
+        )
+
+        # Download INNO Setup installer in background
+        self.downloader = ThreadedDownloader()
+        self.downloader.download(
+            "AppImageTool",
+            f"{APPIMAGE_DOWNLOAD}/appimagetool-{appimage_type}.AppImage",
+        )
+        self.downloader.download(
+            "AppImageRun",
+            f"{APPIMAGE_DOWNLOAD}/AppRun-{appimage_type}",
+        )
+        self.downloader.download(
+            "AppImageRuntime",
+            f"{APPIMAGE_DOWNLOAD}/runtime-{appimage_type}",
+        )
+
+        print()  # newline
+
+    def make_appimage(self):
+        """
+        Make AppImage installers
+        """
+        if self.appimagekitdir is None:
+            return
+
+        print("Making Linux universal packages with AppImage")
+
+        installer_build_dir = self.packaging_dir / "build" / "kithare.AppDir"
+        rmtree(installer_build_dir)  # clean old build dir
+
+        bin_dir = installer_build_dir / "usr" / "bin"
+        bin_dir.mkdir(parents=True)
+
+        # copy dist exe
+        copy(self.exepath, bin_dir)
+
+        # copy desktop file
+        copy(self.packaging_dir / "kithare.desktop", installer_build_dir)
+
+        # copy icon file
+        copy(
+            self.packaging_dir.parents[1] / "assets" / "kithare.png",
+            installer_build_dir,
+        )
+
+        self.appimagekitdir.mkdir(parents=True, exist_ok=True)
+
+        appimagekit = {
+            "AppImageTool": self.appimagekitdir / "appimagetool.AppImage",
+            "AppImageRun": self.appimagekitdir / "AppRun",
+            "AppImageRuntime": self.appimagekitdir / "runtime",
+        }
+
+        if self.downloader is not None:
+            if self.downloader.is_downloading():
+                print("Waiting for AppImageKit downloads to finish")
+
+            for name, data in self.downloader.get_finished():
+                if data is None:
+                    raise BuildError(f"Failed to fetch downloads of {name}")
+
+                # save download in file
+                appimagekit[name].write_bytes(data)
+                appimagekit[name].chmod(0o775)
+                print(f"Successfully downloaded {name}!")
+
+        if self.machine not in {"x86", "x64"}:
+            # A workaround for AppImage bug on arm docker
+            # https://github.com/AppImage/AppImageKit/issues/1056
+            run_cmd(
+                "sed",
+                "-i",
+                r"s|AI\x02|\x00\x00\x00|",
+                appimagekit["AppImageTool"],
+                strict=True,
+            )
+
+        # copy main runfile to AppDir
+        copy(appimagekit["AppImageRun"], installer_build_dir)
+
+        dist_image = (
+            self.packaging_dir
+            / "dist"
+            / f"kithare-{self.version}-{self.appimagekitdir.name}.AppImage"
+        )
+        dist_image.parent.mkdir(exist_ok=True)
+
+        run_cmd(
+            appimagekit["AppImageTool"],
+            "--appimage-extract-and-run",
+            "--runtime-file",
+            appimagekit["AppImageRuntime"],
+            installer_build_dir,
+            dist_image,
+            strict=True,
+        )
+        dist_image.chmod(0o775)
+        print(f"Successfully made AppImage installer in '{dist_image}'!\n")
 
     def debian_package(self):
         """
@@ -156,13 +293,14 @@ class LinuxPackager(Packager):
         """
         print("\nMaking Debian .deb installer")
 
-        version = f"{KITHARE_VERSION}-{VERSION_PACKAGE_REV}"
         machine = convert_machine(self.machine, ConvertType.LINUX_DEB)
 
-        installer_dir = self.packaging_dir / "build" / f"kithare_{version}"
-        rmtree(installer_dir)  # clean old dir
+        installer_dir = (
+            self.packaging_dir / "build" / "Debian" / f"kithare_{self.version}"
+        )
+        rmtree(installer_dir.parent)  # clean old dir
 
-        bin_dir = installer_dir / "usr" / "local" / "bin"
+        bin_dir = installer_dir / "usr" / "bin"
         bin_dir.mkdir(parents=True)
 
         # copy dist exe
@@ -174,7 +312,7 @@ class LinuxPackager(Packager):
 
         control_file = self.packaging_dir / "debian_control.txt"
         write_control_file.write_text(
-            f"Version: {version}\n"
+            f"Version: {self.version}\n"
             + f"Architecture: {machine}\n"
             + control_file.read_text()
         )
@@ -186,7 +324,6 @@ class LinuxPackager(Packager):
         license_file.rename(license_file.with_name("copyright"))
 
         dist_dir = self.packaging_dir / "dist"
-        dist_dir.mkdir(exist_ok=True)
 
         run_cmd("dpkg-deb", "--build", installer_dir, dist_dir, strict=True)
         print(".deb file was made successfully\n")
@@ -199,15 +336,19 @@ class LinuxPackager(Packager):
                 "--to-rpm",
                 "--keep-version",
                 f"--target={rpm_machine}",
-                dist_dir / f"kithare_{version}_{machine}.deb",
+                dist_dir / f"kithare_{self.version}_{machine}.deb",
                 strict=True,
             )
 
-            gen_rpm = self.basepath / f"kithare-{version}.{rpm_machine}.rpm"
+            gen_rpm = (
+                self.packaging_dir.parents[1]
+                / f"kithare-{self.version.replace('-', '_', 1)}.{rpm_machine}.rpm"
+            )
             try:
                 rpm_file = copy(gen_rpm, dist_dir)
             finally:
-                gen_rpm.unlink()
+                if gen_rpm.is_file():
+                    gen_rpm.unlink()
 
             print(f"Generated rpm package in '{rpm_file}'!\n")
 
@@ -216,6 +357,8 @@ class LinuxPackager(Packager):
         Make installer for Linux
         """
         super().package()
+
+        self.make_appimage()
 
         print("Using Linux installer configuration")
         print("Testing for Debian")
@@ -238,7 +381,9 @@ class MacPackager(Packager):
     """
 
 
-def get_packager(basepath: Path, exepath: Path, machine: str, use_alien: bool):
+def get_packager(
+    basepath: Path, exepath: Path, machine: str, version: str, use_alien: bool
+):
     """
     Get appropriate packager class for handling packaging
     """
@@ -247,16 +392,16 @@ def get_packager(basepath: Path, exepath: Path, machine: str, use_alien: bool):
         if use_alien:
             raise BuildError("'--use-alien' is not a supported flag on this OS")
 
-        return WindowsPackager(basepath, exepath, machine)
+        return WindowsPackager(basepath, exepath, machine, version)
 
     if system == "Linux":
-        return LinuxPackager(basepath, exepath, machine, use_alien)
+        return LinuxPackager(basepath, exepath, machine, version, use_alien)
 
     if system == "Darwin":
         if use_alien:
             raise BuildError("'--use-alien' is not a supported flag on this OS")
 
-        return MacPackager(basepath, exepath, machine)
+        return MacPackager(basepath, exepath, machine, version)
 
     raise BuildError(
         "Cannot generate installer as your platform could not be determined!"
